refactor(edgeless): refactor clipboard utility functions and update references:
