+++ packages/blocks/src/page-block/edgeless/utils/bound-utils.ts
@@ -1,0 +1,15 @@
+import { Bound } from '@blocksuite/phasor';
+
+import type { EdgelessElement } from '../../../index.js';
+import type { Selectable } from '../services/tools-manager.js';
+import { isTopLevelBlock } from './query.js';
+
+export function getGridBound(ele: Selectable) {
+  return isTopLevelBlock(ele) ? Bound.deserialize(ele.xywh) : ele.gridBound;
+}
+
+export function edgelessElementsBound(elements: EdgelessElement[]) {
+  return elements.reduce((prev, element) => {
+    return prev.unite(getGridBound(element));
+  }, getGridBound(elements[0]));
+}
