+++ packages/blocks/src/page-block/edgeless/utils/clipboard-utils.ts
@@ -7,18 +7,8 @@ import {
 
 import type { EdgelessElement } from '../../../index.js';
 import type { EdgelessPageBlockComponent } from '../edgeless-page-block.js';
-import type { Selectable } from '../services/tools-manager.js';
+import { edgelessElementsBound, getGridBound } from './bound-utils.js';
 import { isTopLevelBlock } from './query.js';
-
-export function getGridBound(ele: Selectable) {
-  return isTopLevelBlock(ele) ? Bound.deserialize(ele.xywh) : ele.gridBound;
-}
-
-export function edgelessElementsBound(elements: EdgelessElement[]) {
-  return elements.reduce((prev, element) => {
-    return prev.unite(getGridBound(element));
-  }, getGridBound(elements[0]));
-}
 
 export function getCopyElements(
   edgeless: EdgelessPageBlockComponent,
